+++ src/parkingspot/parkingspot.entity.ts
